+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.4
 ### v0.4.0
- - Major refactoring, change to an Annotator `class` to prepare and add annotations
+ - Major refactoring, change to an Annotator `class` to prepare and add annotations distinctly
+ - Support for `violinplot`
  - Fixes in rendering of non significant tests results after multiple comparisons correction
  - Fix the printout of the star annotation legend
  - Many unit and integration tests